Extract view-related methods into their own mixin class.
They were placed in the mixin intended as a base for content, but content
and view are two distinct entities. Mixing the mixins is confusing as they
do not share state.
